Separate all test classes into separate files. Fix Poker Test issues. Add new files to .gitignore.
