+++ .gitignore
@@ -1,1 +1,4 @@
 .DS_Store
+__pycache__
+*.pyc
+.idea
